% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChicagoTemperatures.R
\docType{data}
\name{ChicagoTemperatures}
\alias{ChicagoTemperatures}
\title{Chicago temperature data}
\format{}
\source{

}
\usage{
data(ChicagoTemperature)
}
\description{
Chicago temperature data
}
\references{

}
\keyword{datasets}
