% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coExp.R
\docType{data}
\name{coExp}
\alias{coExp}
\title{Randomly sampled exposure from Colorado counties}
\format{
matrix
}
\source{
\url{https://aqs.epa.gov/aqsweb/airdata/download_files.html}
}
\usage{
data(coExp)
}
\description{
Matrix of five different exposures, each measured over 40 weeks.
}
\references{
\url{https://www.epa.gov/outdoor-air-quality-data}
}
\keyword{datasets}
