% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.models.R
\name{combine.models.tdlmm}
\alias{combine.models.tdlmm}
\title{Combines information from DLMs of mixture exposures.}
\usage{
combine.models.tdlmm(mlist)
}
\arguments{
\item{mlist}{a list of models}
}
\value{
A data frame with model fit information of the models included in the list
}
\description{
Method for combining information from DLMs of mixture exposures.
}
\details{
combine.models.tdlmm
}
