% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cppIntersection}
\alias{cppIntersection}
\title{fast set intersection tool assumes sorted vectors A and B}
\usage{
cppIntersection(A, B)
}
\arguments{
\item{A}{sorted integer vector A}

\item{B}{sorted integer vector B}
}
\value{
vector of resulting intersection
}
\description{
fast set intersection tool assumes sorted vectors A and B
}
