% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{ctrl.family}
\alias{ctrl.family}
\title{Family control settings for dlmtree model fitting}
\usage{
ctrl.family(binomial.size = 1, formula.zi = NULL)
}
\arguments{
\item{binomial.size}{integer type scalar (if all equal, default: 1) or vector defining binomial size for 'logit' family.}

\item{formula.zi}{(only applies to family = 'zinb') object of class formula, a symbolic description of the fixed effect of
zero-inflated (ZI) model to be fitted, e.g. y ~ a + b. This only applies to ZINB where covariates for
ZI model are different from NB model. This is set to the argument 'formula' by default.}
}
\value{
A list of family control parameters.
}
\description{
Family control settings for dlmtree model fitting
}
