% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{ctrl.mcmc}
\alias{ctrl.mcmc}
\title{MCMC control settings for dlmtree model fitting}
\usage{
ctrl.mcmc(n.trees = 20, n.burn = 1000, n.iter = 2000, n.thin = 10)
}
\arguments{
\item{n.trees}{integer for number of trees in ensemble.}

\item{n.burn}{integer for length of MCMC burn-in.}

\item{n.iter}{integer for number of MCMC iterations to run model after burn-in.}

\item{n.thin}{integer MCMC thinning factor, i.e. keep every tenth iteration.}
}
\value{
A list of MCMC control parameters.
}
\description{
MCMC control settings for dlmtree model fitting
}
