% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlmEst}
\alias{dlmEst}
\title{Calculates the distributed lag effect with DLM matrix for linear models.}
\usage{
dlmEst(dlm, nlags, nsamp)
}
\arguments{
\item{dlm}{A numeric matrix containing the model fit information}

\item{nlags}{total number of lags}

\item{nsamp}{number of mcmc iterations}
}
\value{
A cube object of lag effect x lag x mcmc
}
\description{
Calculates the distributed lag effect with DLM matrix for linear models.
}
