% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmtree-package.R
\docType{package}
\name{dlmtree-package}
\alias{dlmtree-package}
\title{dlmtree: Bayesian Treed Distributed Lag Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Estimate and interpretation of treed distributed lag models (DLMs). The methods estimate constrained DLMs using the Bayesian additive regression trees (BART) framework. Includes various extensions of treed DLMs. These include: treed DLMs and distributed lag mixture models as described in Mork and Wilson (2023) \doi{ 10.1111/biom.13568}; treed distributed lag nonlinear models as described in Mork and Wilson (2021) \doi{ 10.1093/biostatistics/kxaa051}; heterogeneous DLMs as described in Mork et al. (2024) \doi{10.1080/01621459.2023.2258595}. The package also includes a Shiny App to help interpret results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danielmork/dlmtree}
  \item \url{https://danielmork.github.io/dlmtree/}
  \item Report bugs at \url{https://github.com/danielmork/dlmtree/issues}
}

}
\author{
\strong{Maintainer}: Daniel Mork \email{dmork@hsph.harvard.edu} (\href{https://orcid.org/0000-0002-7924-0706}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Seongwon Im \email{Seongwon.Im@colostate.edu} (\href{https://orcid.org/0009-0000-8447-5852}{ORCID})
  \item Ander Wilson \email{Ander.Wilson@colostate.edu} (\href{https://orcid.org/0000-0003-4774-3883}{ORCID})
}

}
\keyword{internal}
