% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{dlmtree.control.tdlnm}
\alias{dlmtree.control.tdlnm}
\title{Control settings for dlmtree model fitting, when used for TDLNM}
\usage{
dlmtree.control.tdlnm(
  exposure.splits = 20,
  time.split.prob = NULL,
  exposure.se = NULL
)
}
\arguments{
\item{exposure.splits}{scalar indicating the number of splits (divided
evenly across quantiles of the exposure data) or list with two components:
'type' = 'values' or 'quantiles', and 'split.vals' = a numerical
vector indicating the corresponding exposure values or quantiles for splits.}

\item{time.split.prob}{probability vector of a spliting probabilities for time lags. (default: uniform probabilities)}

\item{exposure.se}{numerical matrix of exposure standard errors with same
size as exposure.data or a scalar smoothing factor representing a uniform
smoothing factor applied to each exposure measurement. (default: sd(exposure.data)/2)}
}
\value{
list of TDLNM control parameters.
}
\description{
Control settings for dlmtree model fitting, when used for TDLNM
}
