% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlmtreeHDLMMGaussian}
\alias{dlmtreeHDLMMGaussian}
\title{dlmtree model with HDLMM approach}
\usage{
dlmtreeHDLMMGaussian(model)
}
\arguments{
\item{model}{A list of parameter and data contained for the model fitting}
}
\value{
A list of dlmtree model fit, mainly posterior mcmc samples
}
\description{
dlmtree model with HDLMM approach
}
