% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlmtreeShared}
\alias{dlmtreeShared}
\title{dlmtree model with shared HDLM approach}
\usage{
dlmtreeShared(model)
}
\arguments{
\item{model}{A list of parameter and data contained for the model fitting}
}
\value{
A list of dlmtree model fit, mainly posterior mcmc samples
}
\description{
dlmtree model with shared HDLM approach
}
