% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlmtreeTDLM_cpp}
\alias{dlmtreeTDLM_cpp}
\title{dlmtree model with nested HDLM approach}
\usage{
dlmtreeTDLM_cpp(model)
}
\arguments{
\item{model}{A list of parameter and data contained for the model fitting}
}
\value{
A list of dlmtree model fit, mainly posterior mcmc samples
}
\description{
dlmtree model with nested HDLM approach
}
