% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlmtreeTDLMFixedGaussian}
\alias{dlmtreeTDLMFixedGaussian}
\title{dlmtree model with fixed Gaussian approach}
\usage{
dlmtreeTDLMFixedGaussian(model)
}
\arguments{
\item{model}{A list of parameter and data contained for the model fitting}
}
\value{
A list of dlmtree model fit, mainly posterior mcmc samples
}
\description{
dlmtree model with fixed Gaussian approach
}
