% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlnmMod.R
\name{dlnmMod}
\alias{dlnmMod}
\title{dlnmMod}
\usage{
dlnmMod(X, SE = NULL, X.bins)
}
\arguments{
\item{X}{Exposure data}

\item{SE}{Standard error of exposure data, can be left null}

\item{X.bins}{Interger value specifying the number of evenly spaced quantiles
to split exposures into (0 indicates to use all values as possible splits)
or list("type" = c("values", quantiles"), split.vals = numeric vector)}
}
\value{
Environment with objects and functions
}
\description{
Modifiers environment for DLNM
}
