% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlnmNode.R
\name{dlnmNode}
\alias{dlnmNode}
\title{dlnmNode}
\usage{
dlnmNode(
  depth = 0L,
  gen2 = FALSE,
  xmin = -Inf,
  xmax = Inf,
  tmin = 1L,
  tmax = Inf,
  xsplit = NULL,
  tsplit = NULL,
  counts = c(),
  logprule = 0,
  avail.mod = NULL,
  parent = NULL,
  sib = NULL,
  c1 = NULL,
  c2 = NULL
)
}
\arguments{
\item{depth}{Depth of node in tree (starts at 0)}

\item{gen2}{T/F, does node have child nodes that are both terminal}

\item{xmin}{Min exposure value}

\item{xmax}{Max exposure value}

\item{tmin}{Min time}

\item{tmax}{Max time}

\item{xsplit}{Split on exposure value}

\item{tsplit}{Split on time}

\item{counts}{Number of observations that are in node}

\item{logprule}{Log probability of split}

\item{avail.mod}{Numeric list of available modifiers}

\item{parent}{Reference to parent node}

\item{sib}{Reference to sibling node}

\item{c1}{Reference to child node 1}

\item{c2}{Reference to child node 2}
}
\value{
function environment with objects and methods
}
\description{
DLNM node environment
}
