% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlnmPLEst}
\alias{dlnmPLEst}
\title{Calculates the distributed lag effect with DLM matrix for non-linear models.}
\usage{
dlnmPLEst(dlnm, predAt, nlags, nsamp, center)
}
\arguments{
\item{dlnm}{A numeric matrix containing the model fit information}

\item{predAt}{Number of splits in the model}

\item{nlags}{total number of lags}

\item{nsamp}{number of mcmc iterations}

\item{center}{center parameter}
}
\value{
A cube object of lag effect x lag x mcmc
}
\description{
Calculates the distributed lag effect with DLM matrix for non-linear models.
}
