% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlnmTreeMCMC.R
\name{dlnmTreeMCMC}
\alias{dlnmTreeMCMC}
\title{dlnmTreeMCMC}
\usage{
dlnmTreeMCMC(model, ctr, dgn)
}
\arguments{
\item{model}{Environment with model data and control parameters}

\item{ctr}{Environment with MCMC control parameters}

\item{dgn}{Environment for storing output and diagnostics}
}
\value{
invisible()
}
\description{
Updates all trees in model, alters environments, no return.
}
