% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlnmTreeUpdate.R
\name{dlnmTreeUpdate}
\alias{dlnmTreeUpdate}
\title{dlnmTreeUpdate}
\usage{
dlnmTreeUpdate(tree, step, model)
}
\arguments{
\item{tree}{Reference to tree environment}

\item{step}{Step (1-4 relates to grow, prune, change, swap)}

\item{model}{Model environment}
}
\value{
List detailing update
}
\description{
Propose an update for dlnmTree
}
