% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawTree.R
\name{drawTree}
\alias{drawTree}
\title{drawTree}
\usage{
drawTree(depth, alpha, beta)
}
\arguments{
\item{depth}{depth of a tree}

\item{alpha}{tree shape parameter, 0 < alpha < 1}

\item{beta}{tree size parameter, beta > 0}
}
\value{
A integer value of number of terminal nodes
}
\description{
A recursive function to draw a new tree structure
}
