% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposureCov.R
\docType{data}
\name{exposureCov}
\alias{exposureCov}
\title{Exposure covariance structure}
\format{
matrix
}
\source{
\url{https://aqs.epa.gov/aqsweb/airdata/download_files.html}
}
\usage{
data(exposureCov)
}
\description{
Matrix containing pairwise covariances for real exposure data
consisting of five different exposures, each measured over 37 weeks.
}
\references{
\url{https://www.epa.gov/outdoor-air-quality-data}
}
\keyword{datasets}
