% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdjacencyMatrix.R
\name{getAdjacencyMatrix}
\alias{getAdjacencyMatrix}
\title{Get an adjacency matrix}
\usage{
getAdjacencyMatrix(fips, state_abbv = "CO")
}
\arguments{
\item{...}{FIPS coordinates of counties}
}
\value{
An adjacency matrix consisting of 1s and 0s
}
\description{
get.adjacency.matrix
}
