% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdjacentCounties.R
\name{getAdjacentCounties}
\alias{getAdjacentCounties}
\title{Finding an adjacent county}
\usage{
getAdjacentCounties(fips, state_abbv = "CO")
}
\arguments{
\item{...}{NA}
}
\value{
An adjacent county to a provided county with FIPS coordinate
}
\description{
get.adjacent.counties
}
