% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logPSplit.R
\name{logPSplit}
\alias{logPSplit}
\title{logPSplit}
\usage{
logPSplit(alpha, beta, node.depth, terminal = F)
}
\arguments{
\item{alpha}{Parameter for splitting probability}

\item{beta}{Parameter for splitting probability}

\item{node.depth}{Parameter for splitting probability}

\item{terminal}{Is node terminal}
}
\value{
log probability of split, or log probability of terminal
}
\description{
logPSplit
}
