% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixEst}
\alias{mixEst}
\title{Calculates the lagged interaction effects with MIX matrix for linear models.}
\usage{
mixEst(dlm, nlags, nsamp)
}
\arguments{
\item{dlm}{A numeric matrix containing the model fit information}

\item{nlags}{total number of lags}

\item{nsamp}{number of mcmc iterations}
}
\value{
A cube object of interaction effect x lag x mcmc
}
\description{
Calculates the lagged interaction effects with MIX matrix for linear models.
}
