% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{monotdlnm_Cpp}
\alias{monotdlnm_Cpp}
\title{dlmtree model with monotone tdlnm approach}
\usage{
monotdlnm_Cpp(model)
}
\arguments{
\item{model}{A list of parameter and data contained for the model fitting}
}
\value{
A list of dlmtree model fit, mainly posterior mcmc samples
}
\description{
dlmtree model with monotone tdlnm approach
}
