% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.summary.monotone.R
\name{plot.summary.monotone}
\alias{plot.summary.monotone}
\title{Returns variety of plots for model summary of class 'monotone'}
\usage{
\method{plot}{summary.monotone}(x, plot.type = "mean", val = c(), time = c(), ...)
}
\arguments{
\item{x}{object of class 'summary.monotone', output of summary of 'monotone'}

\item{plot.type}{string indicating plot type, options are 'mean' (default)
which shows mean exposure-time response surface, 'se', 'ci-min', 'ci-max',
'slice' which takes a slice of the plot at a given 'val' or 'time',
'animate' which creates a animation of slices of the surface plot across
exposure values (requires package gganimate)}

\item{val}{exposure value for slice plot}

\item{time}{time value for slice plot}

\item{...}{additional parameters to alter plots: 'main', 'xlab', 'ylab',
'flab' which sets the effect label for surface plots,
'start.time' which sets the first time value}
}
\value{
A plot of distributed lag effect estimated with monotone-TDLNM
}
\description{
Method for returning variety of plots for model summary of class 'monotone'
}
\details{
plot.summary.monotone
}
