% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.summary.tdlm.R
\name{plot.summary.tdlm}
\alias{plot.summary.tdlm}
\title{plot.summary.tdlnm}
\usage{
\method{plot}{summary.tdlm}(x, trueDLM = NULL, ...)
}
\arguments{
\item{x}{object of class 'summary.tdlnm', output of summary of 'tdlnm'}

\item{trueDLM}{A vector of true effects that can be obtained from the simulated data. Only applicable for simulation studies}

\item{...}{additional plotting parameters for title and labels
'flab' which sets the effect label for surface plots,
'start.time' which sets the first time value}
}
\description{
plot.summary.tdlnm
}
