% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.summary.tdlnm.R
\name{plot.summary.tdlnm}
\alias{plot.summary.tdlnm}
\title{plot.summary.tdlnm}
\usage{
\method{plot}{summary.tdlnm}(summary, ptype = "mean", var = c(), lag = c(), ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
plot.summary.tdlnm
}
