% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppRange.R
\name{ppRange}
\alias{ppRange}
\title{Makes a 'pretty' output of a group of numbers}
\usage{
ppRange(r)
}
\arguments{
\item{r}{set of integers to make 'pretty'}
}
\value{
character string of values representing 'r'
}
\description{
Method for making a 'pretty' output of a group of numbers. For
example: 2,3,4,5,8,9,12,15,16 becomes 2-5,8-9,12,15-16
}
\details{
ppRange
}
