% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.dlmtree.R
\name{predict.dlmtree}
\alias{predict.dlmtree}
\title{predict.dlmtree}
\usage{
\method{predict}{dlmtree}(
  object,
  new.data,
  new.exposure.data,
  ...,
  ci.level = 0.95,
  type = "response",
  outcome = NULL,
  fixed.idx = list(),
  est.dlm = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{fitted dlmtree model with dlmType hdlm/hdlmm}

\item{new.data}{new data frame which contains the same covariates and modifiers used to fit HDLM/HDLMM model}

\item{new.exposure.data}{new data frame/list which contains the same length of exposure lags used to fit HDLM/HDLMM model}

\item{ci.level}{credible interval level for posterior predictive distribution}

\item{type}{type of prediction: "response" (default) or "waic". "waic" must be specified with `outcome` parameter}

\item{outcome}{outcome required for WAIC calculation}

\item{fixed.idx}{fixed index}

\item{est.dlm}{flag for estimating dlm effect}

\item{verbose}{true (default) or false: print output}
}
\value{
Posterior predictive distribution draws
}
\description{
Calculate predicted response for HDLM & HDLMM
}
