% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dlmtree.R
\name{print.hdlm}
\alias{print.hdlm}
\title{Print a hdlm Object}
\usage{
\method{print}{hdlm}(x, ...)
}
\arguments{
\item{x}{An object of class hdlm.}

\item{...}{Not used.}
}
\value{
Assorted model output.
}
\description{
Print a hdlm Object
}
