% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dlmtree.R
\name{print.hdlmm}
\alias{print.hdlmm}
\title{Print a hdlmm Object}
\usage{
\method{print}{hdlmm}(x, ...)
}
\arguments{
\item{x}{An object of class hdlmm.}

\item{...}{Not used.}
}
\value{
Assorted model output.
}
\description{
Print a hdlmm Object
}
