% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dlmtree.R
\name{print.monotone}
\alias{print.monotone}
\title{Print a monotone Object}
\usage{
\method{print}{monotone}(x, ...)
}
\arguments{
\item{x}{An object of class monotone}

\item{...}{Not used.}
}
\value{
Assorted model output.
}
\description{
Print a monotone Object
}
