% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/print.dlmtree.R
\name{print}
\alias{print}
\alias{print.tdlnm}
\alias{print.tdlm}
\alias{print.tdlmm}
\alias{print.hdlm}
\alias{print.hdlmm}
\alias{print.monotone}
\title{print}
\usage{
print(x, ...)

\method{print}{tdlnm}(x, ...)

\method{print}{tdlm}(x, ...)

\method{print}{tdlmm}(x, ...)

\method{print}{hdlm}(x, ...)

\method{print}{hdlmm}(x, ...)

\method{print}{monotone}(x, ...)
}
\arguments{
\item{x}{an object of class 'tdlm', 'tdlmm', 'tdlnm', 'hdlm', 'hdlmm', 'monotone'}

\item{...}{not used.}
}
\value{
assorted model output including model formula call and available methods
}
\description{
print generic function for S3method
}
