% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.hdlm.R
\name{print.summary.hdlm}
\alias{print.summary.hdlm}
\title{print.summary.hdlm}
\usage{
\method{print}{summary.hdlm}(object, digits = 4, cw.only = TRUE)
}
\arguments{
\item{object}{an object of type 'summary.hdlm', result of call to summary.hdlm()}

\item{digits}{integer number of digits to round}

\item{cw.only}{print only results for exposures with critical windows}
}
\value{
output in R console
}
\description{
print.summary.hdlm
}
