% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.hdlmm.R
\name{print.summary.hdlmm}
\alias{print.summary.hdlmm}
\title{print.summary.hdlmm}
\usage{
\method{print}{summary.hdlmm}(object, digits = 4, cw.only = TRUE)
}
\arguments{
\item{object}{an object of type 'summary.hdlmm', result of call to summary.hdlmm()}

\item{digits}{integer number of digits to round}

\item{cw.only}{print only results for exposures with critical windows}
}
\value{
output in R console
}
\description{
print.summary.hdlmm
}
