% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.summary}
\alias{print.summary}
\title{print.summary}
\usage{
\method{print}{summary}(x, digits = 3, ...)
}
\arguments{
\item{x}{a summary object resulting from summary() applied to an object of 
class 'tdlm', 'tdlmm', 'tdlnm', 'hdlm', 'hdlmm', 'monotone'}

\item{digits}{number of decimal places to round the numeric values to}

\item{...}{additional parameters}
}
\value{
summary output of a model fit printed in the R console
}
\description{
print.summary generic function for S3method
}
