% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.tdlmm.R
\name{print.summary.tdlmm}
\alias{print.summary.tdlmm}
\title{print.summary.tdlmm}
\usage{
\method{print}{summary.tdlmm}(object, digits = 4, cw.only = TRUE)
}
\arguments{
\item{object}{an object of type 'summary.tdlmm', result of call to summary.tdlmm()}

\item{digits}{integer number of digits to round}

\item{cw.only}{print only results for exposures with critical windows}
}
\value{
output in R console
}
\description{
print.summary.tdlmm
}
