% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dlmtree.R
\name{print.tdlm}
\alias{print.tdlm}
\title{Print a tdlm Object}
\usage{
\method{print}{tdlm}(x, ...)
}
\arguments{
\item{x}{An object of class tdlm.}

\item{...}{Not used.}
}
\value{
Assorted model output.
}
\description{
Print a tdlm Object
}
