% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dlmtree.R
\name{print.tdlmm}
\alias{print.tdlmm}
\title{Print a tdlmm Object}
\usage{
\method{print}{tdlmm}(x, ...)
}
\arguments{
\item{x}{An object of class tdlmm.}

\item{...}{Not used.}
}
\value{
Assorted model output.
}
\description{
Print a tdlmm Object
}
