% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_pgdraw}
\alias{rcpp_pgdraw}
\title{Multiple draw polya gamma latent variable for var c[i] with size b[i]}
\usage{
rcpp_pgdraw(b, z)
}
\arguments{
\item{b}{vector of binomial sizes}

\item{z}{vector of parameters}
}
\value{
Eigen::VectorXd
}
\description{
Multiple draw polya gamma latent variable for var c[i] with size b[i]
}
