% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtmvnorm}
\alias{rtmvnorm}
\title{truncated multivariate normal sampler, mean mu, cov sigma, truncated (0, Inf)}
\usage{
rtmvnorm(mu, sigma, iter)
}
\arguments{
\item{mu}{vector of mean parameters}

\item{sigma}{covariance matrix}

\item{iter}{number of iterations}
}
\value{
VectorXd
}
\description{
truncated multivariate normal sampler, mean mu, cov sigma, truncated (0, Inf)
}
