% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estDLM.R
\name{ruleIdx}
\alias{ruleIdx}
\title{ruleIdx}
\usage{
ruleIdx(mod, mem.safe = FALSE)
}
\arguments{
\item{mod}{a list of modifier splitting rules}

\item{mem.safe}{boolean memory parameter}
}
\value{
A list of weights per rule with modifiers
}
\description{
A function to calculate the weights for each rule
}
