% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbd_dlmtree.R
\docType{data}
\name{sbd}
\alias{sbd}
\title{Simulated covariates and exposures in the context of association 
between birth weight and maternal exposure to environmental chemicals}
\format{
data.frame
}
\source{
\url{https://aqs.epa.gov/aqsweb/airdata/download_files.html}
}
\usage{
data(sbd_dlmtree)
}
\description{
data frame of maternal covariates, birth information, 
and five exposures, each measured over 37 weeks.
}
\references{
\url{https://www.epa.gov/outdoor-air-quality-data}
}
\keyword{datasets}
