% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbd_dlmtree.R
\docType{data}
\name{sbd_dlmtree}
\alias{sbd_dlmtree}
\title{Simulated covariates and exposures in the context of association 
between birth weight and maternal exposure to environmental chemicals}
\format{
data.frame;
}
\usage{
data(sbd_dlmtree)
}
\description{
data frame of maternal covariates, birth information, 
and five exposures, each measured over 37 weeks.
}
\keyword{datasets}
