% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleModelMatrix.R
\name{scaleModelMatrix}
\alias{scaleModelMatrix}
\title{scaleModelMatrix}
\usage{
scaleModelMatrix(M)
}
\arguments{
\item{M}{a matrix to center and scale}
}
\value{
a scaled matrix
}
\description{
A function to center and scale a matrix
}
