% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{shiny}
\alias{shiny}
\title{shiny}
\usage{
shiny(fit)
}
\arguments{
\item{fit}{an object of class hdlm or hdlmm to which S3method is applied}
}
\value{
A 'shiny' interface for further analysis on heterogeneous analyses. The interface includes tabs for modifier selection, personalized exposure effects and subgroup-specific effects.
}
\description{
shiny generic function for S3method
}
