% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.tdlmm.zinb.R
\name{sim.tdlmm.zinb}
\alias{sim.tdlmm.zinb}
\title{sim.tdlmm.zinb}
\usage{
sim.tdlmm.zinb(
  sim = 1,
  ctnum = 20,
  week = 561,
  expList = NULL,
  data_zinb = NULL,
  rho = 0.95,
  spTau = 1
)
}
\arguments{
\item{sim}{integer (1-3) specifying simulation scenario}

\item{ctnum}{number of counties (or spatial units)}

\item{week}{Weeks for each county}

\item{expList}{Named list of exposure data}

\item{data_zinb}{data}
}
\value{

}
\description{
Simulation scenarios to accompany TDLM/TDLMM for count data
}
