% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.tdlnm.R
\name{sim.tdlnm}
\alias{sim.tdlnm}
\title{sim.tdlnm}
\usage{
sim.tdlnm(sim = "A", error.to.signal = 1)
}
\arguments{
\item{sim}{character (A - D) specifying simulation scenario}

\item{error.to.signal}{scalar value setting error: sigma^2/var(f)}
}
\value{
Simulated data and true parameters
}
\description{
sim.tdlnm
}
\details{
Simulation scenarios:
\itemize{
\item Scenario A: Piecewise constant effect
\item Scenario B: Linear effect
\item Scenario C: Logistic effect, piecewise in time
\item Scenario D: Logistic effect, smooth in time
}
}
\examples{
sim.tdlnm(sim = "A", error.to.signal = 1)

}
