% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPairwise.R
\name{spatial_pairwise}
\alias{spatial_pairwise}
\title{Spatial adjacency matrix to two nodes vectors}
\usage{
spatial_pairwise(AdjMatrix)
}
\arguments{
\item{...}{Adjacency matrix}
}
\value{
Two nodes vectors
}
\description{
spatial_pairwise
}
