% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{splitPIP}
\alias{splitPIP}
\title{Calculates the posterior includsion probability (PIP).}
\usage{
splitPIP(dlnm, nlags, niter)
}
\arguments{
\item{dlnm}{A numeric matrix containing the model fit information}

\item{nlags}{total number of lags}

\item{niter}{number of mcmc iterations}
}
\value{
A matrix of split counts per mcmc
}
\description{
Calculates the posterior includsion probability (PIP).
}
