% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitpoints.R
\name{splitpoints}
\alias{splitpoints}
\title{Determine split points for continuous modifiers}
\usage{
splitpoints(object, var, round = NULL)
}
\arguments{
\item{object}{An object of class dlmtree.}

\item{var}{The name of a continuous variable for which the split points will be reported.}

\item{round}{The number of decimal places to round the variable (var) to. No rounding occurs if round=NULL (default). For positive integer values of round, the variable will be rounded and split points will be reported at the resulting level.}
}
\value{
A data.frame with split points and the probability that a split point was >= that split point value.
}
\description{
Determine split points for continuous modifiers
}
