% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.monotone.R
\name{summary.monotone}
\alias{summary.monotone}
\title{summary.monotone}
\usage{
\method{summary}{monotone}(
  object,
  pred.at = NULL,
  cenval = 0,
  conf.level = 0.95,
  exposure.se = NULL,
  mcmc = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{an object of class 'monotone'}

\item{pred.at}{numerical vector of exposure values to make predictions for
at each time period}

\item{cenval}{scalar exposure value that acts as a reference point for
predictions at all other exposure values}

\item{conf.level}{confidence level for computation of credible intervals}

\item{exposure.se}{scalar smoothing factor, if different from model}

\item{mcmc}{true or false (default): return MCMC samplers}

\item{verbose}{true (default) or false: print output}

\item{...}{additional parameters}
}
\value{
Summary of monotone fit
}
\description{
summary.monotone
}
