% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tdlm.R
\name{summary.tdlm}
\alias{summary.tdlm}
\title{Creates a summary object of class 'tdlm'}
\usage{
\method{summary}{tdlm}(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{an object of dlm class 'tdlm' (i.e. a linear effect DLM)}

\item{conf.level}{confidence level for computation of credible intervals}

\item{...}{additional parameters}
}
\value{
list of type 'summary.tdlm'
}
\description{
Method for creating a summary object of class 'tdlm'
}
\details{
summary.tdlm
}
