% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tdlmm.R
\name{summary.tdlmm}
\alias{summary.tdlmm}
\title{summary.tdlmm}
\usage{
\method{summary}{tdlmm}(
  object,
  conf.level = 0.95,
  marginalize = "mean",
  log10BF.crit = 0.5,
  verbose = TRUE,
  keep.mcmc = FALSE,
  ...
)
}
\arguments{
\item{object}{an object of type 'tdlmm', the output from tdlmm()}

\item{conf.level}{confidence level (default = 0.95)}

\item{marginalize}{value(s) for calculating marginal DLMs, defaults to
"mean", can also specify a percentile from 1-99 for all other exposures, or
a named vector with specific values for each exposure}

\item{log10BF.crit}{Bayes Factor criteria for selecting exposures and
interactions, such that log10(BayesFactor) > x. Default = 0.5}

\item{verbose}{show progress in console}

\item{keep.mcmc}{keep all mcmc iterations (large memory requirement)}

\item{...}{additional parameters}
}
\value{
list of type 'summary.tdlmm'
}
\description{
summary.tdlmm
}
