% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tdlmm_Cpp}
\alias{tdlmm_Cpp}
\title{dlmtree model with tdlmm approach}
\usage{
tdlmm_Cpp(model)
}
\arguments{
\item{model}{A list of parameter and data contained for the model fitting}
}
\value{
A list of dlmtree model fit, mainly posterior mcmc samples
}
\description{
dlmtree model with tdlmm approach
}
