% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdlnm.gaussian.R
\name{tdlnm.gaussian}
\alias{tdlnm.gaussian}
\title{tdlnm.gaussian}
\usage{
tdlnm.gaussian(model)
}
\arguments{
\item{model}{Environment with model data and control parameters}
}
\value{
invisible()
}
\description{
Sets up and runs MCMC for tdlnm model with Gaussian errors.
Modifies environment, no return.
}
