% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdlnm.R
\name{tdlnm}
\alias{tdlnm}
\title{tdlnm}
\usage{
tdlnm(
  formula,
  data,
  exposure.data,
  exposure.se = NULL,
  exposure.splits = 100,
  n.trees = 20,
  n.burn = 2000,
  n.iter = 5000,
  n.thin = 10,
  tree.params = c(0.95, 2),
  step.prob = c(0.3, 0.4),
  subset = c(),
  verbose = TRUE,
  diagnostics = FALSE,
  ...
)
}
\arguments{
\item{formula}{object of class formula, a symbolic description of the fixed 
effect model to be fitted, e.g. y ~ a + b}

\item{data}{data frome containing variables used in the formula}

\item{exposure.data}{numerical matrix of complete exposure data with same
length as data}

\item{exposure.se}{numerical matrix of exposure standard errors with same
length as data or scalar smoothing factor representing a uniform standard
error over each exposure measurement}

\item{exposure.splits}{scalar indicating the number of splits (divided
evenly across quantiles of the exposure data) or list with two components:
'type' set to 'values' or 'quantiles', and 'split.vals' set as a numerical
vector indicating the corresponding exposure values or quantiles for splits}

\item{n.trees}{integer for number of trees in ensemble (default = 20)}

\item{n.burn}{integer for length of burn-in}

\item{n.iter}{integer for number of iterations to run model after burn-in}

\item{n.thin}{integer thinning factor, i.e. keep every tenth iteration}

\item{tree.params}{numerical vector of alpha and beta hyperparameters 
controlling tree depth (see Bayesian CART, 1998)}

\item{step.prob}{numerical vector for probability of 1) grow/prune, 2) change}

\item{subset}{integer vector to analyze only a subset of data and exposures}

\item{verbose}{true/false print output}

\item{diagnostics}{true/false keep model diagnostic such as terminal nodes, 
acceptance details, etc.}

\item{...}{NA}
}
\value{

}
\description{
tdlnm
}
