% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdlnm.sim.R
\name{tdlnm.sim}
\alias{tdlnm.sim}
\title{tdlnm.sim}
\usage{
tdlnm.sim(effect = "A", error.to.signal = 1)
}
\arguments{
\item{effect}{character (A-D) indicating simulation scenario}

\item{error.to.signal}{scalar value setting error: sigma^2/var(f)}
}
\value{

}
\description{
tdlnm.sim
}
