% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{zeroToInfNormCDF}
\alias{zeroToInfNormCDF}
\title{Integrates (0,inf) over multivariate normal}
\usage{
zeroToInfNormCDF(mu, sigma)
}
\arguments{
\item{mu}{vector of mean parameters}

\item{sigma}{covariance matrix}
}
\value{
double
}
\description{
Integrates (0,inf) over multivariate normal
}
