% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbCo.R
\docType{data}
\name{zinbCo}
\alias{zinbCo}
\title{Time-series exposure data for ZINB simulated data}
\format{
data.frame;
}
\source{
\url{https://aqs.epa.gov/aqsweb/airdata/download_files.html}
}
\usage{
data(zinbCo)
}
\description{
Data.frame containing a sample of weekly average wildfire PM, PM2.5, O3
across a range of counties of Colorado. The exposure data was downloaded from US EPA
(https://aqs.epa.gov/aqsweb/airdata/download_files.html) daily data summaries
and averaged by week.
}
\references{
\url{https://www.epa.gov/outdoor-air-quality-data}
}
\keyword{datasets}
