% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.models.R
\name{combine.models}
\alias{combine.models}
\title{combine.models}
\usage{
combine.models(mlist)
}
\arguments{
\item{mlist}{a list of models}
}
\value{
A data frame with model fit information of the models included in the list
}
\description{
Method for combining information from DLMs of single exposure
}
