% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{diagnose}
\alias{diagnose}
\title{diagnose}
\usage{
diagnose(x, ...)
}
\arguments{
\item{x}{a summary object resulting from summary() applied to an object of 
class 'tdlm', 'tdlmm', 'tdlnm', 'hdlm', 'hdlmm', 'monotone'}

\item{...}{not used.}
}
\value{
shiny interface for assessing model convergence. The interface includes 
tabs for MCMC diagnostics such as trace plots, density plots, and convergence measures 
for distributed lag effects, DLM tree sizes, and hyperparameters.
}
\description{
diagnose generic function for S3method
}
