% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmtree-package.R
\docType{package}
\name{dlmtree-package}
\alias{dlmtree-package}
\title{dlmtree: Bayesian Treed Distributed Lag Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Estimation of distributed lag models (DLMs) based on a Bayesian additive regression trees framework. Includes several extensions of DLMs: treed DLMs and distributed lag mixture models (Mork and Wilson, 2023) \doi{10.1111/biom.13568}; treed distributed lag nonlinear models (Mork and Wilson, 2022) \doi{10.1093/biostatistics/kxaa051}; heterogeneous DLMs (Mork, et. al., 2024) \doi{10.1080/01621459.2023.2258595}; monotone DLMs (Mork and Wilson, 2024) \doi{10.1214/23-BA1412}. The package also includes visualization tools and a 'shiny' interface to help interpret results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danielmork/dlmtree}
  \item \url{https://danielmork.github.io/dlmtree/}
  \item Report bugs at \url{https://github.com/danielmork/dlmtree/issues}
}

}
\author{
\strong{Maintainer}: Daniel Mork \email{dmork@hsph.harvard.edu} (\href{https://orcid.org/0000-0002-7924-0706}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Seongwon Im \email{Seongwon.Im@colostate.edu} (\href{https://orcid.org/0009-0000-8447-5852}{ORCID})
  \item Ander Wilson \email{Ander.Wilson@colostate.edu} (\href{https://orcid.org/0000-0003-4774-3883}{ORCID})
}

}
\keyword{internal}
