% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip.R
\name{pip}
\alias{pip}
\title{Posterior inclusion probabilities (PIPs) for modifiers in HDLM}
\usage{
pip(object, type = 1)
}
\arguments{
\item{object}{An object of class dlmtree.}

\item{type}{Type=1 indicates single modifier PIPs. Type=2 indicates joint modifier PIPs for two modifiers.}
}
\value{
A vector (type=1) or data.frame (type=2) of PIPs.
}
\description{
Posterior inclusion probabilities (PIPs) for modifiers in HDLM
}
