% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.summary.tdlm.R
\name{plot.summary.tdlm}
\alias{plot.summary.tdlm}
\title{Plots a distributed lag function for model summary of 'tdlm'}
\usage{
\method{plot}{summary.tdlm}(x, trueDLM = NULL, ...)
}
\arguments{
\item{x}{object of class 'summary.tdlm', output of summary of 'tdlm'}

\item{trueDLM}{A vector of true effects that can be obtained from the simulated data. Only applicable for simulation studies}

\item{...}{additional plotting parameters for title and labels
'start.time' which sets the first time value}
}
\value{
A plot of distributed lag effect estimated with tdlm
}
\description{
Method for plotting a distributed lag function for model summary of 'tdlm'
}
\details{
plot.summary.tdlm
}
\examples{
D <- sim.tdlmm(sim = "A", mean.p = 0.5, n = 1000)
fit <- dlmtree(y ~ .,
               data = D$dat,
               exposure.data = D$exposures[[1]],
               dlm.type = "linear",
               family = "logit",
               binomial.size = 1)
fit_sum <- summary(fit)
plot(fit_sum)

}
