% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.summary.tdlmm.R
\name{plot.summary.tdlmm}
\alias{plot.summary.tdlmm}
\title{plot.summary.tdlmm}
\usage{
\method{plot}{summary.tdlmm}(
  object,
  type = "marginal",
  exposure1 = NULL,
  exposure2 = NULL,
  time1 = c(),
  time2 = c(),
  show.cw = TRUE,
  cw.plots.only = TRUE,
  trueDLM = NULL,
  scale = NULL,
  ...
)
}
\arguments{
\item{object}{object of type 'summary.tdlmm' from summary.tdlmm() output}

\item{type}{plot type, 'marginal' (default)}

\item{exposure1}{exposure for plotting DLM}

\item{exposure2}{exposure paired with 'exposure1' for plotting interaction}

\item{time1}{plot a cross section from an interaction plot at specific time
for 'exposure1'}

\item{time2}{plot a cross section from an interaction plot at specific time
for 'exposure2'}

\item{show.cw}{indicate location of critical windows in interaction plot
with red points}

\item{cw.plots.only}{show only plots with critical windows}

\item{...}{additional plotting parameters for manipulating title and labels}
}
\value{
ggplot
}
\description{
Method for plotting DLMMs from summary of tdlmm. Includes plots
for marginal exposure effects as well as interactions between two exposures.
}
