% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.summary.tdlnm.R
\name{plot.summary.tdlnm}
\alias{plot.summary.tdlnm}
\title{plot.summary.tdlnm}
\usage{
\method{plot}{summary.tdlnm}(summary, plot.type = "mean", val = c(), time = c(), trueDLM = NULL, ...)
}
\arguments{
\item{summary}{object of class 'summary.tdlnm', output of summary of 'tdlnm'}

\item{plot.type}{string indicating plot type, options are 'mean' (default)
which shows mean exposure-time response surface, 'se', 'ci-min', 'ci-max',
'slice' which takes a slice of the plot at a given 'val' or 'time',
'animate' which creates a animation of slices of the surface plot across
exposure values (requires package gganimate)}

\item{val}{exposure value for slice plot}

\item{time}{time value for slice plot}

\item{trueDLM}{A vector of true effects that can be obtained from the simulated data. Only applicable for simulation studies}

\item{...}{additional parameters to alter plots: 'main', 'xlab', 'ylab',
'flab' which sets the effect label for surface plots,
'start.time' which sets the first time value}
}
\description{
plot.summary.tdlnm
}
