% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.summary.tdlnm.R
\name{plot.summary.tdlnm}
\alias{plot.summary.tdlnm}
\title{Returns variety of plots for model summary of class 'tdlnm'}
\usage{
\method{plot}{summary.tdlnm}(x, plot.type = "mean", val = c(), time = c(), trueDLM = NULL, ...)
}
\arguments{
\item{x}{object of class 'summary.tdlnm', output of summary of 'tdlnm'}

\item{plot.type}{string indicating plot type, options are 'mean' (default)
which shows mean exposure-time response surface, 'se', 'ci-min', 'ci-max',
'slice' which takes a slice of the plot at a given 'val' or 'time',
'animate' which creates a animation of slices of the surface plot across
exposure values (requires package gganimate)}

\item{val}{exposure value for slice plot}

\item{time}{time value for slice plot}

\item{trueDLM}{A vector of true effects that can be obtained from the simulated data. Only applicable for simulation studies}

\item{...}{additional plotting parameters for title and labels
'flab' which sets the effect label for surface plots,
'start.time' which sets the first time value}
}
\value{
A plot of distributed lag effect estimated with tdlnm
}
\description{
Method for returning variety of plots for model summary of class 'tdlnm'
}
\details{
plot.summary.tdlnm
}
\examples{
D <- sim.tdlnm(sim = "A", error.to.signal = 1)
fit <- dlmtree(formula = y ~ .,
               data = D$dat,
               exposure.data = D$exposures,
               dlm.type = "nonlinear",
               family = "gaussian")
fit_sum <- summary(fit)
plot(fit_sum)

}
