% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm25Exposures.R
\docType{data}
\name{pm25Exposures}
\alias{pm25Exposures}
\title{PM2.5 Exposure data}
\format{

}
\source{

}
\usage{
data(pm25Exposures)
}
\description{
PM2.5 Exposure data
}
\references{

}
\keyword{datasets}
