% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dlmtree.R
\name{print.tdlnm}
\alias{print.tdlnm}
\alias{print.tdlm}
\alias{print.tdlmm}
\alias{print.hdlm}
\alias{print.hdlmm}
\alias{print.monotone}
\title{Print method for model class 'tdlm', 'tdlmm', 'tdlnm', 'hdlm', 'hdlmm', 'monotone'}
\usage{
\method{print}{tdlnm}(x, ...)

\method{print}{tdlm}(x, ...)

\method{print}{tdlmm}(x, ...)

\method{print}{hdlm}(x, ...)

\method{print}{hdlmm}(x, ...)

\method{print}{monotone}(x, ...)
}
\arguments{
\item{x}{An object of class monotone}

\item{...}{Not used.}
}
\value{
Assorted model output.
}
\description{
Print method for model class 'tdlm', 'tdlmm', 'tdlnm', 'hdlm', 'hdlmm', 'monotone'
}
