% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.hdlm.R
\name{print.summary.hdlm}
\alias{print.summary.hdlm}
\title{Prints an overview with summary of model class 'hdlm'}
\usage{
\method{print}{summary.hdlm}(x, digits = 3, cw.only = TRUE, ...)
}
\arguments{
\item{x}{an object of type 'summary.hdlm', result of call to summary.hdlm()}

\item{digits}{integer number of digits to round}

\item{cw.only}{print only results for exposures with critical windows}

\item{...}{additional parameters}
}
\value{
output of hdlm fit in R console
}
\description{
Method for printing an overview with summary of model class 'hdlm'
}
\details{
print.summary.hdlm
}
\examples{
D <- sim.hdlmm(sim = "B", n = 1000)
fit <- dlmtree(y ~ ., 
               data = D$dat,
               exposure.data = D$exposures,
               dlm.type = "linear",
               family = "gaussian",
               het = TRUE)
fit_sum <- summary(fit)
print(fit_sum)

}
