% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.hdlmm.R
\name{print.summary.hdlmm}
\alias{print.summary.hdlmm}
\title{Prints an overview with summary of model class 'hdlmm'}
\usage{
\method{print}{summary.hdlmm}(x, digits = 3, cw.only = TRUE, ...)
}
\arguments{
\item{x}{an object of type 'summary.hdlmm', result of call to summary.hdlmm()}

\item{digits}{integer number of digits to round}

\item{cw.only}{print only results for exposures with critical windows}

\item{...}{additional parameters}
}
\value{
output of hdlmm fit in R console
}
\description{
Method for printing an overview with summary of model class 'hdlmm'
}
\details{
print.summary.hdlmm
}
\examples{
\donttest{
D <- sim.hdlmm(sim = "D", n = 1000)
fit <- dlmtree(y ~ ., 
               data = D$dat,
               exposure.data = D$exposures,
               dlm.type = "linear",
               family = "gaussian",
               mixture = TRUE,
               het = TRUE)
fit_sum <- summary(fit)
print(fit_sum)
}

}
