% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.monotone.R
\name{print.summary.monotone}
\alias{print.summary.monotone}
\title{print.summary.monotone}
\usage{
\method{print}{summary.monotone}(x, digits = 3, ...)
}
\arguments{
\item{x}{an object of type 'summary.monotone', result of call to summary.monotone()}

\item{digits}{integer number of digits to round}

\item{...}{additional parameters}
}
\value{
output in R console
}
\description{
print.summary.monotone
}
