% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.monotone.R
\name{print.summary.monotone}
\alias{print.summary.monotone}
\title{Prints an overview with summary of model class 'monotone'}
\usage{
\method{print}{summary.monotone}(x, digits = 3, ...)
}
\arguments{
\item{x}{an object of type 'summary.monotone', result of call to summary.monotone()}

\item{digits}{integer number of digits to round}

\item{...}{additional parameters}
}
\value{
output in R console
}
\description{
Method for printing an overview with summary of model class 'monotone'
}
\details{
print.summary.monotone
}
\examples{
D <- sim.tdlnm(sim = "A", error.to.signal = 1)
fit <- dlmtree(formula = y ~ .,
               data = D$dat,
               exposure.data = D$exposures,
               dlm.type = "monotone",
               family = "gaussian")
fit_sum <- summary(fit)
print(fit_sum)

}
