% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/print.summary.hdlm.R,
%   R/print.summary.hdlmm.R, R/print.summary.monotone.R, R/print.summary.tdlm.R,
%   R/print.summary.tdlmm.R, R/print.summary.tdlnm.R
\name{print.summary}
\alias{print.summary}
\alias{print.summary.hdlm}
\alias{print.summary.hdlmm}
\alias{print.summary.monotone}
\alias{print.summary.tdlm}
\alias{print.summary.tdlmm}
\alias{print.summary.tdlnm}
\title{print.summary}
\usage{
\method{print}{summary}(x, digits = 3, ...)

\method{print.summary}{hdlm}(x, digits = 3, ...)

\method{print.summary}{hdlmm}(x, digits = 3, ...)

\method{print.summary}{monotone}(x, digits = 3, ...)

\method{print.summary}{tdlm}(x, digits = 3, ...)

\method{print.summary}{tdlmm}(x, digits = 3, ...)

\method{print.summary}{tdlnm}(x, digits = 3, ...)
}
\arguments{
\item{x}{a summary object resulting from summary() applied to an object of 
class 'tdlm', 'tdlmm', 'tdlnm', 'hdlm', 'hdlmm', 'monotone'}

\item{digits}{number of decimal places to round the numeric values to}

\item{...}{additional parameters}
}
\value{
summary output of a model fit printed in the R console
}
\description{
print.summary generic function for S3method
}
