% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.tdlm.R
\name{print.summary.tdlm}
\alias{print.summary.tdlm}
\title{print.summary.tdlm}
\usage{
\method{print}{summary.tdlm}(object, digits = 3)
}
\arguments{
\item{object}{an object of type 'summary.tdlnm', result of call to summary.tdlnm()}

\item{digits}{integer number of digits to round}
}
\value{
output in R console
}
\description{
print.summary.tdlm
}
