% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.tdlm.R
\name{print.summary.tdlm}
\alias{print.summary.tdlm}
\title{Prints an overview with summary of model class 'tdlm'}
\usage{
\method{print}{summary.tdlm}(x, digits = 3, ...)
}
\arguments{
\item{x}{an object of type 'summary.tdlnm', result of call to summary.tdlnm()}

\item{digits}{integer number of digits to round}

\item{...}{additional parameters}
}
\value{
output of tdlm fit in R console
}
\description{
Method for printing an overview with summary of model class 'tdlm'
}
\details{
print.summary.tdlm
}
\examples{
D <- sim.tdlmm(sim = "A", mean.p = 0.5, n = 1000)
fit <- dlmtree(y ~ .,
               data = D$dat,
               exposure.data = D$exposures[[1]],
               dlm.type = "linear",
               family = "logit",
               binomial.size = 1)
fit_sum <- summary(fit)
print(fit_sum)

}
