% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleModelMatrix.R
\name{scaleModelMatrix}
\alias{scaleModelMatrix}
\title{Scale model matrix}
\usage{
scaleModelMatrix(M)
}
\arguments{
\item{M}{Matrix of predictors}
}
\value{
Matrix of scaled predictors
}
\description{
Scale model matrix
}
