% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleModelMatrix.R
\name{scaleModelMatrix}
\alias{scaleModelMatrix}
\title{Centers and scales a matrix}
\usage{
scaleModelMatrix(M)
}
\arguments{
\item{M}{a matrix to center and scale}
}
\value{
a scaled matrix
}
\description{
Method for centering and scaling a matrix
}
\details{
scaleModelMatrix
}
