% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.hdlm.R
\name{shiny.hdlm}
\alias{shiny.hdlm}
\title{shiny.hdlm}
\usage{
\method{shiny}{hdlm}(fit)
}
\arguments{
\item{fit}{an object of class 'hdlm'}
}
\value{
A shinyapp interface
}
\description{
A function to execute a shinyApp to provide comprehensive analysis with HDLM. The shinyApp includes PIP, split points, individualized & subgroup-specific effect.
}
