% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.hdlm.R
\name{shiny.hdlm}
\alias{shiny.hdlm}
\title{Executes a 'shiny' app for HDLM.}
\usage{
\method{shiny}{hdlm}(fit)
}
\arguments{
\item{fit}{an object of class 'hdlm'}
}
\value{
A 'shiny' app interface
}
\description{
Method for executing a 'shiny' app to provide comprehensive analysis with HDLM. The app includes PIP, split points, individualized & subgroup-specific effects.
}
\details{
shiny.hdlm
}
