% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.hdlmm.R
\name{shiny.hdlmm}
\alias{shiny.hdlmm}
\title{Executes a 'shiny' app for HDLMM.}
\usage{
\method{shiny}{hdlmm}(fit)
}
\arguments{
\item{fit}{an object of class 'hdlmm'}
}
\value{
A 'shiny' app interface
}
\description{
Method for executing a 'shiny' app to provide comprehensive analysis with HDLMM. The app includes PIP, split points, individualized & subgroup-specific effects for exposure of interest.
}
\details{
shiny.hdlmm
}
