% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny}
\alias{shiny}
\title{shiny}
\usage{
shiny(x, ...)
}
\arguments{
\item{x}{an object to which S3method is applied}

\item{...}{additional parameters}
}
\description{
shiny generic function for S3method
}
