% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/shiny.hdlm.R, R/shiny.hdlmm.R
\name{shiny}
\alias{shiny}
\alias{shiny.hdlm}
\alias{shiny.hdlmm}
\title{shiny}
\usage{
shiny(fit)

\method{shiny}{hdlm}(fit)

\method{shiny}{hdlmm}(fit)
}
\arguments{
\item{fit}{an object of class 'hdlm', 'hdlmm' to which S3method is applied}
}
\value{
'shiny' interface for further analysis on heterogeneous analyses. 
The interface includes tabs for modifier selection, personalized exposure 
effects and subgroup-specific effects.
}
\description{
shiny generic function for S3method
}
