% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.tdlmm.R
\name{sim.tdlmm}
\alias{sim.tdlmm}
\title{sim.tdlmm}
\usage{
sim.tdlmm(
  sim = "A",
  error = 10,
  mean.p = 0.5,
  n.exp = 25,
  prop.active = 0.05,
  n = 5000,
  expList = NULL,
  ctnum = 20,
  week = 561,
  data_zinb = NULL,
  r = 1
)
}
\arguments{
\item{sim}{character (A - F) specifying simulation scenario}

\item{error}{positive scalar specifying error variance}

\item{mean.p}{scalar between zero and one specifying mean probability
for simulation scenario one}

\item{n.exp}{number of exposures for simulation scenarios three and four}

\item{prop.active}{proportion of active exposures for simulation scenario
three}

\item{n}{sample size for simulation}

\item{expList}{named list of exposure data}

\item{ctnum}{number of counties}

\item{week}{number of weeks for each county, must be between 1-561}

\item{data_zinb}{covariate data}

\item{r}{dispersion parameter}

\item{effect.size}{the effect size of the window of susceptibility, resulting in controlling the magnitude of the count outcome}
}
\value{
Simulated data and true parameters
}
\description{
Simulation scenarios to accompany TDLM/TDLMM
}
