% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.tdlmm.zinb.R
\name{sim.tdlmm.zinb}
\alias{sim.tdlmm.zinb}
\title{sim.tdlmm.zinb}
\usage{
sim.tdlmm.zinb(
  sim = 1,
  ctnum = 20,
  week = 561,
  expList = NULL,
  data_zinb = NULL,
  effect.size = 0.1,
  r = 1
)
}
\arguments{
\item{sim}{integer (1-2) specifying simulation scenario}

\item{ctnum}{number of counties}

\item{week}{number of weeks for each county, must be between 1-561}

\item{expList}{named list of exposure data}

\item{data_zinb}{covariate data}

\item{effect.size}{the effect size of the window of susceptibility, resulting in controlling the magnitude of the count outcome}

\item{r}{dispersion parameter}
}
\value{
A simulated dataset, true parameter values, zero/non-zero proportions
}
\description{
Simulation scenarios to accompany TDLM/TDLMM for count data
}
