% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.tdlmm.zinb.R
\name{sim.tdlmm.zinb}
\alias{sim.tdlmm.zinb}
\title{sim.tdlmm.zinb}
\usage{
sim.tdlmm.zinb(
  sim = 1,
  ctnum = 20,
  week = 561,
  expList = NULL,
  data_zinb = NULL,
  zi_mean = 0,
  effect.size = 0.1,
  r = 1
)
}
\arguments{
\item{sim}{integer (1-2) specifying simulation scenario}

\item{ctnum}{number of counties,}

\item{week}{Weeks for each county, must be between 1-561}

\item{expList}{Named list of exposure data}

\item{data_zinb}{covariate data}

\item{zi_mean}{Mean of at-risk coefficients to control the zero-inflation of the count outcome}

\item{effect.size}{Parameter to control the magnitude of the count outcome}

\item{r}{Dispersion parameter}
}
\value{
A simulated dataset with true parameter values & zero-inflation information
}
\description{
Simulation scenarios to accompany TDLM/TDLMM for count data
}
