% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.tdlnm.R
\name{sim.tdlnm}
\alias{sim.tdlnm}
\title{sim.tdlnm}
\usage{
sim.tdlnm(effect = "A", error.to.signal = 1)
}
\arguments{
\item{effect}{character (A - D) indicating simulation scenario}

\item{error.to.signal}{scalar value setting error: sigma^2/var(f)}
}
\description{
sim.tdlnm
}
