% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.hdlm.R
\name{summary.hdlm}
\alias{summary.hdlm}
\title{summary.hdlm}
\usage{
\method{summary}{hdlm}(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{an object of class 'hdlm'}

\item{conf.level}{confidence level for computation of credible intervals}

\item{...}{additional parameters}
}
\value{
list of type 'summary.hdlm'
}
\description{
summary.hdlm
}
\examples{
D <- sim.hdlmm(sim = "B", n = 1000)
fit <- dlmtree(y ~ ., 
               data = D$dat,
               exposure.data = D$exposures,
               dlm.type = "linear",
               family = "gaussian",
               het = TRUE)
summary(fit)

}
