% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.hdlmm.R
\name{summary.hdlmm}
\alias{summary.hdlmm}
\title{Creates a summary object of class 'hdlmm'}
\usage{
\method{summary}{hdlmm}(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{an object of class 'hdlmm'}

\item{conf.level}{confidence level for computation of credible intervals}

\item{...}{additional parameters}
}
\value{
list of type 'summary.hdlmm'
}
\description{
Method for creating a summary object of class 'hdlm'
}
\details{
summary.hdlmm
}
\examples{
D <- sim.hdlmm(sim = "D", n = 1000)
fit <- dlmtree(y ~ ., 
               data = D$dat,
               exposure.data = D$exposures,
               dlm.type = "linear",
               family = "gaussian",
               mixture = TRUE,
               het = TRUE)
summary(fit)

}
