% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.hdlm.R, R/summary.hdlmm.R,
%   R/summary.monotone.R, R/summary.tdlm.R, R/summary.tdlmm.R, R/summary.tdlnm.R
\name{summary.hdlm}
\alias{summary.hdlm}
\alias{summary.hdlmm}
\alias{summary.monotone}
\alias{summary.tdlm}
\alias{summary.tdlmm}
\alias{summary.tdlnm}
\title{Summary method for model class 'tdlm', 'tdlmm', 'tdlnm', 'hdlm', 'hdlmm', 'monotone'}
\usage{
\method{summary}{hdlm}(object, conf.level = 0.95, ...)

\method{summary}{hdlmm}(object, conf.level = 0.95, ...)

\method{summary}{monotone}(
  object,
  pred.at = NULL,
  cenval = 0,
  conf.level = 0.95,
  exposure.se = NULL,
  mcmc = FALSE,
  verbose = TRUE,
  ...
)

\method{summary}{tdlm}(object, conf.level = 0.95, ...)

\method{summary}{tdlmm}(
  object,
  conf.level = 0.95,
  marginalize = "mean",
  log10BF.crit = 0.5,
  verbose = TRUE,
  keep.mcmc = FALSE,
  ...
)

\method{summary}{tdlnm}(
  object,
  pred.at = NULL,
  cenval = 0,
  conf.level = 0.95,
  exposure.se = NULL,
  mcmc = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{an object of class 'tdlnm', result of a call to tdlnm}

\item{conf.level}{confidence level for computation of credible intervals}

\item{...}{additional parameters}

\item{pred.at}{numerical vector of exposure values to make predictions for
at each time period}

\item{cenval}{scalar exposure value that acts as a reference point for
predictions at all other exposure values}

\item{exposure.se}{scalar smoothing factor, if different from model}

\item{mcmc}{TRUE or FALSE (default): return MCMC samplers}

\item{verbose}{TRUE (default) or FALSE: print output}

\item{marginalize}{value(s) for calculating marginal DLMs, defaults to
"mean", can also specify a percentile from 1-99 for all other exposures, or
a named vector with specific values for each exposure}

\item{log10BF.crit}{Bayes Factor criteria for selecting exposures and
interactions, such that log10(BayesFactor) > x. Default = 0.5}

\item{keep.mcmc}{keep all mcmc iterations (large memory requirement)}
}
\value{
list object for model class 'summary.tdlm', 'summary.tdlmm', 'summary.tdlnm', 'summary.hdlm', 'summary.hdlmm', 'summary.monotone'
}
\description{
Summary method for model class 'tdlm', 'tdlmm', 'tdlnm', 'hdlm', 'hdlmm', 'monotone'
}
