% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tdlm.R
\name{summary.tdlm}
\alias{summary.tdlm}
\title{summary.tdlm}
\usage{
\method{summary}{tdlm}(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{an object of dlm class 'tdlm' (i.e. a linear effect DLM)}

\item{conf.level}{confidence level for computation of credible intervals}

\item{...}{additional parameters}
}
\value{
list of type 'summary.tdlm'
}
\description{
summary.tdlm
}
\examples{
D <- sim.tdlmm(sim = "A", mean.p = 0.5, n = 1000)
fit <- dlmtree(y ~ ., 
               data = D$dat, 
               exposure.data = D$exposures[[1]],
               dlm.type = "linear",
               family = "logit",
               binomial.size = 1)
summary(fit)

}
