% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tdlm.R
\name{summary.tdlm}
\alias{summary.tdlm}
\title{summary.tdlm}
\usage{
\method{summary}{tdlm}(object, conf.level = 0.95)
}
\arguments{
\item{object}{an object of class 'tdlm', result of a call to tdlnm using
zero splits (i.e. a linear effect DLM)}

\item{conf.level}{confidence level for computation of credible intervals}
}
\value{
list of type 'summary.tdlm'
}
\description{
summary.tdlm
}
