% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tdlnm.R
\name{summary.tdlnm}
\alias{summary.tdlnm}
\title{summary.tdlnm}
\usage{
\method{summary}{tdlnm}(
  object,
  pred.at = NULL,
  cenval = 0,
  conf.level = 0.95,
  exposure.se = NULL,
  zirt.cond = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{an object of class 'tdlnm', result of a call to tdlnm}

\item{pred.at}{numerical vector of exposure values to make predictions for
at each time period}

\item{cenval}{scalar exposure value that acts as a reference point for
predictions at all other exposure values}

\item{conf.level}{confidence level for computation of credible intervals}

\item{exposure.se}{scalar smoothing factor, if different from model}
}
\value{

}
\description{
summary.tdlnm
}
